+++ src/lib/Warp.NET/Content/Conversion/Blobs/BlobContentConverter.cs
@@ -1,6 +1,6 @@
 namespace Warp.NET.Content.Conversion.Blobs;
 
-public sealed class BlobContentConverter : IContentConverter<BlobBinary>
+internal sealed class BlobContentConverter : IContentConverter<BlobBinary>
 {
 	public static BlobBinary Construct(string inputPath)
 	{