+++ src/lib/Warp.NET/Content/Conversion/Sounds/SoundBinary.cs
@@ -1,6 +1,6 @@
 namespace Warp.NET.Content.Conversion.Sounds;
 
-public record SoundBinary(short Channels, int SampleRate, short BitsPerSample, byte[] Data) : IBinary<SoundBinary>
+internal record SoundBinary(short Channels, int SampleRate, short BitsPerSample, byte[] Data) : IBinary<SoundBinary>
 {
 	public ContentType ContentType => ContentType.Sound;