+++ src/lib/Warp.NET/Content/Entity.cs
@@ -1,0 +1,3 @@
+namespace Warp.NET.Content;
+
+public record Entity(Dictionary<string, string> Properties, List<Brush> Brushes);