+++ src/lib/Warp.NET/Content/Face.cs
@@ -1,0 +1,3 @@
+namespace Warp.NET.Content;
+
+public record Face(Vector3 P1, Vector3 P2, Vector3 P3, string TextureName, Plane TextureAxisU, Plane TextureAxisV, Vector2 TextureScale);